Sidebar refactor
