+++ client/src/preview/route/digitaltwins/editor/sidebarFunctions.ts
@@ -2,28 +2,19 @@ import { addOrUpdateFile, FileState } from 'preview/store/file.slice';
 import DigitalTwin from 'preview/util/digitalTwin';
 import { Dispatch, SetStateAction } from 'react';
 import { useDispatch } from 'react-redux';
-import { TreeItem, TreeItemProps } from '@mui/x-tree-view/TreeItem';
-import * as React from 'react';
 import LibraryAsset from 'preview/util/libraryAsset';
 import {
   addOrUpdateLibraryFile,
   LibraryConfigFile,
 } from 'preview/store/libraryConfigFiles.slice';
-
-export const getFileTypeFromExtension = (fileName: string): string => {
-  const extension = fileName.split('.').pop()?.toLowerCase();
-  if (extension === 'md') return 'description';
-  if (extension === 'json' || extension === 'yaml' || extension === 'yml')
-    return 'config';
-  return 'lifecycle';
-};
-
-export const fetchData = async (digitalTwin: DigitalTwin) => {
-  await digitalTwin.getDescriptionFiles();
-  await digitalTwin.getLifecycleFiles();
-  await digitalTwin.getConfigFiles();
-  await digitalTwin.getAssetFiles();
-};
+import {
+  getFileTypeFromExtension,
+  updateFileState,
+} from 'preview/util/fileUtils';
+import {
+  fetchAndSetFileContent,
+  fetchAndSetFileLibraryContent,
+} from './sidebarFetchers';
 
 export const handleFileClick = (
   fileName: string,
@@ -70,103 +61,6 @@ export const handleFileClick = (
     );
   }
 };
-
-export const renderFileTreeItems = (
-  label: string,
-  filesToRender: string[],
-  asset: DigitalTwin | LibraryAsset,
-  setFileName: Dispatch<SetStateAction<string>>,
-  setFileContent: Dispatch<SetStateAction<string>>,
-  setFileType: Dispatch<SetStateAction<string>>,
-  files: FileState[],
-  tab: string,
-  dispatch: ReturnType<typeof useDispatch>,
-  setIsLibraryFile: Dispatch<SetStateAction<boolean>>,
-  setLibraryAssetPath: Dispatch<SetStateAction<string>>,
-  library?: boolean,
-  libraryFiles?: LibraryConfigFile[],
-  assetPath?: string,
-) => (
-  <TreeItem
-    itemId={`${label.toLowerCase()}-${label}`}
-    label={label as TreeItemProps['label']}
-  >
-    {filesToRender.map((item) => (
-      <TreeItem
-        key={item}
-        itemId={`${label.toLowerCase()}-${item}`}
-        label={item}
-        onClick={() =>
-          handleFileClick(
-            item,
-            asset!,
-            setFileName,
-            setFileContent,
-            setFileType,
-            files,
-            tab,
-            setIsLibraryFile,
-            setLibraryAssetPath,
-            dispatch,
-            library || undefined,
-            libraryFiles || undefined,
-            assetPath || undefined,
-          )
-        }
-      />
-    ))}
-  </TreeItem>
-);
-
-export const getFilteredFileNames = (type: string, files: FileState[]) =>
-  files
-    .filter(
-      (file) => file.isNew && getFileTypeFromExtension(file.name) === type,
-    )
-    .map((file) => file.name);
-
-export const renderFileSection = (
-  label: string,
-  type: string,
-  filesToRender: string[],
-  asset: DigitalTwin | LibraryAsset,
-  setFileName: Dispatch<SetStateAction<string>>,
-  setFileContent: Dispatch<SetStateAction<string>>,
-  setFileType: Dispatch<SetStateAction<string>>,
-  files: FileState[],
-  tab: string,
-  dispatch: ReturnType<typeof useDispatch>,
-  setIsLibraryFile: Dispatch<SetStateAction<boolean>>,
-  setLibraryAssetPath: Dispatch<SetStateAction<string>>,
-  library?: boolean,
-  fileLibrary?: LibraryConfigFile[],
-) => (
-  <TreeItem itemId={`${label.toLowerCase()}-${label}`} label={label}>
-    {filesToRender.map((item) => (
-      <TreeItem
-        key={item}
-        itemId={`${label.toLowerCase()}-${item}`}
-        label={item}
-        onClick={() =>
-          handleFileClick(
-            item,
-            asset!,
-            setFileName,
-            setFileContent,
-            setFileType,
-            files,
-            tab,
-            setIsLibraryFile,
-            setLibraryAssetPath,
-            dispatch,
-            library || undefined,
-            fileLibrary || undefined,
-          )
-        }
-      />
-    ))}
-  </TreeItem>
-);
 
 export const handleCreateFileClick = (
   fileName: string,
@@ -301,94 +195,6 @@ export const handleReconfigureFileClick = async (
       setLibraryAssetPath!(assetPath!);
     }
   }
-};
-
-export const fetchAndSetFileContent = async (
-  fileName: string,
-  digitalTwin: DigitalTwin | null,
-  setFileName: Dispatch<SetStateAction<string>>,
-  setFileContent: Dispatch<SetStateAction<string>>,
-  setFileType: Dispatch<SetStateAction<string>>,
-  library?: boolean,
-  assetPath?: string,
-) => {
-  try {
-    let fileContent;
-    if (library) {
-      fileContent = await digitalTwin!.DTAssets.getLibraryFileContent(
-        assetPath!,
-        fileName,
-      );
-    } else {
-      fileContent = await digitalTwin!.DTAssets.getFileContent(fileName);
-    }
-    if (fileContent) {
-      updateFileState(
-        fileName,
-        fileContent,
-        setFileName,
-        setFileContent,
-        setFileType,
-      );
-    }
-  } catch {
-    setFileContent(`Error fetching ${fileName} content`);
-  }
-};
-
-export const fetchAndSetFileLibraryContent = async (
-  fileName: string,
-  libraryAsset: LibraryAsset | null,
-  setFileName: Dispatch<SetStateAction<string>>,
-  setFileContent: Dispatch<SetStateAction<string>>,
-  setFileType: Dispatch<SetStateAction<string>>,
-  isNew: boolean,
-  setIsLibraryFile: Dispatch<SetStateAction<boolean>>,
-  setLibraryAssetPath: Dispatch<SetStateAction<string>>,
-  dispatch?: ReturnType<typeof useDispatch>,
-) => {
-  try {
-    const fileContent = await libraryAsset!.libraryManager.getFileContent(
-      libraryAsset!.isPrivate,
-      libraryAsset!.path,
-      fileName,
-    );
-
-    dispatch!(
-      addOrUpdateLibraryFile({
-        assetPath: libraryAsset!.path,
-        fileName,
-        fileContent,
-        isNew,
-        isModified: false,
-      }),
-    );
-    if (fileContent) {
-      updateFileState(
-        fileName,
-        fileContent,
-        setFileName,
-        setFileContent,
-        setFileType,
-      );
-    }
-    setIsLibraryFile(true);
-    setLibraryAssetPath(libraryAsset!.path);
-  } catch {
-    setFileContent(`Error fetching ${fileName} content`);
-  }
-};
-
-export const updateFileState = (
-  fileName: string,
-  fileContent: string,
-  setFileName: Dispatch<SetStateAction<string>>,
-  setFileContent: Dispatch<SetStateAction<string>>,
-  setFileType: Dispatch<SetStateAction<string>>,
-) => {
-  setFileName(fileName);
-  setFileContent(fileContent);
-  setFileType(fileName.split('.').pop()!);
 };
 
 export const handleAddFileClick = (
